server-side/requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.0.0
cryptography==43.0.0
Cython==3.0.10
dl==0.1.0
docutils==0.21.2
filelock==3.15.4
Flask==3.0.3
Flask_Cors==4.0.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==7.8.2
jnius==1.1.0
keyring==25.2.1
numpy==2.0.1
Pillow==10.4.0
protobuf==5.27.2
pyodbc==5.1.0
pyOpenSSL==24.2.1
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.0.7
Sphinx==7.4.7
thread==2.0.5
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
trove_classifiers==2024.7.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1